requirements.txt
# ── Web server ────────────────────────────────────────────────
fastapi==0.115.0
uvicorn==0.30.6

# ── HTTP clients ──────────────────────────────────────────────
httpx==0.27.2
requests==2.32.3

# ── LLM SDKs ──────────────────────────────────────────────────
openai==1.51.0
anthropic==0.34.2

# ── Observability ─────────────────────────────────────────────
langfuse==2.53.2

# ── Data ──────────────────────────────────────────────────────
pandas==2.2.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# ── Templating ────────────────────────────────────────────────
jinja2==3.1.4

# ── Scraping ──────────────────────────────────────────────────
playwright==1.47.0
beautifulsoup4==4.12.3
lxml==5.3.0

# ── NLP / ML ──────────────────────────────────────────────────
scikit-learn>=1.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0

# ── Testing + stats ───────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
scipy==1.14.1

# ── Utilities ─────────────────────────────────────────────────
python-dateutil==2.9.0
pytz==2024.2
tenacity==9.0.0